feat: refactor App component and add new UI elements for better organization and styling
