+++ src/App.css
@@ -1,0 +1,66 @@
+.app-root {
+  position: relative;
+  width: 100%;
+  height: 100%;
+}
+
+.app-debug-ui {
+  position: absolute;
+  left: 16px;
+  top: 72px;
+  max-width: 440px;
+  z-index: 2;
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+}
+
+.app-alive-panel {
+  position: absolute;
+  top: 16px;
+  right: 16px;
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+  align-items: flex-end;
+}
+
+.app-alive-title {
+  font-weight: 600;
+  font-size: 14px;
+}
+
+.app-victory-overlay {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  padding: 24px 32px;
+  border-radius: 12px;
+  background: rgba(12, 14, 32, 0.85);
+  color: #f7f8ff;
+  text-align: center;
+  box-shadow: 0 12px 32px rgba(0, 0, 0, 0.45);
+}
+
+.app-victory-title {
+  margin: 0;
+  font-size: 22px;
+}
+
+.app-victory-subtitle {
+  margin: 12px 0 0 0;
+}
+
+.app-victory-button {
+  margin-top: 16px;
+}
+
+.app-controls {
+  position: absolute;
+  bottom: 16px;
+  left: 50%;
+  transform: translateX(-50%);
+  display: flex;
+  gap: 12px;
+}
